refactor(compiler): extract migration dependencies
