+++ packages/compiler/src/schema/migration-diff.ts
@@ -1,4 +1,4 @@
-import { canonicalBytes, compareAscii, digest } from "../canonical";
+import { canonicalBytes, compareAscii } from "../canonical";
 import type {
 	MigrationPlanV1,
 	MigrationStepKindV1,
@@ -12,6 +12,11 @@ import {
 } from "./migration-classification";
 import type { MigrationClassification } from "./migration-classification";
 import {
+	expandReferencedKeyDependencies,
+	type MigrationStepDependency,
+} from "./migration-dependencies";
+import { createMigrationStep as step } from "./migration-step";
+import {
 	childRecords,
 	mapByIdentity,
 	mapIdentityBackward,
@@ -20,24 +25,6 @@ import {
 } from "./projection";
 
 type JsonRecord = Readonly<Record<string, unknown>>;
-
-function step(
-	input: Readonly<{
-		kind: MigrationStepKindV1;
-		targetIdentity: string;
-		containerIdentity: string;
-		lock: MigrationStepV1["lock"];
-		scansData: boolean;
-		rewritesTable: boolean;
-		reversibleWithoutData: boolean;
-		classification: MigrationClassification;
-	}>,
-): MigrationStepV1 {
-	return {
-		stepId: digest("questpie-migration-step-v1", input),
-		...input,
-	};
-}
 
 const kindRank: readonly MigrationStepKindV1[] = [
 	"createApplicationSchema",
@@ -62,10 +49,7 @@ const kindRank: readonly MigrationStepKindV1[] = [
 function sortSteps(
 	steps: MigrationStepV1[],
 	renames: MigrationPlanV1["renames"] = [],
-	dependencies: readonly Readonly<{
-		predecessorStepId: string;
-		dependentStepId: string;
-	}>[] = [],
+	dependencies: readonly MigrationStepDependency[] = [],
 ): MigrationStepV1[] {
 	const replacements: Readonly<Record<string, MigrationStepKindV1>> = {
 		addConstraint: "dropConstraint",
@@ -467,183 +451,11 @@ export function destructiveDeltaSteps(
 				}),
 			);
 	}
-	const dependencies: Array<{
-		predecessorStepId: string;
-		dependentStepId: string;
-	}> = [];
-	const stepByKindAndIdentity = new Map(
-		steps.map((candidate) => [
-			`${candidate.kind}\0${candidate.targetIdentity}`,
-			candidate,
-		]),
-	);
-	const ensureStep = (candidate: MigrationStepV1): MigrationStepV1 => {
-		const key = `${candidate.kind}\0${candidate.targetIdentity}`;
-		const existing = stepByKindAndIdentity.get(key);
-		if (existing) return existing;
-		steps.push(candidate);
-		stepByKindAndIdentity.set(key, candidate);
-		return candidate;
-	};
-	const depend = (predecessor: MigrationStepV1, dependent: MigrationStepV1) =>
-		dependencies.push({
-			predecessorStepId: predecessor.stepId,
-			dependentStepId: dependent.stepId,
-		});
-
-	for (const droppedConstraint of steps.filter(
-		(candidate) => candidate.kind === "dropConstraint",
-	)) {
-		const referencedCollection = baseCollections.get(
-			droppedConstraint.containerIdentity,
-		);
-		const constraint = referencedCollection
-			? childRecords(referencedCollection, "constraints").find(
-					(candidate) =>
-						candidate.identity === droppedConstraint.targetIdentity,
-				)
-			: undefined;
-		if (
-			!constraint ||
-			(constraint.kind !== "primaryKey" && constraint.kind !== "unique")
-		)
-			continue;
-
-		const replacementConstraint = stepByKindAndIdentity.get(
-			`addConstraint\0${mapIdentityForward(droppedConstraint.targetIdentity, renames)}`,
-		);
-		for (const sourceCollection of base.collections) {
-			for (const relation of childRecords(sourceCollection, "relations")) {
-				if (
-					relation.target !== droppedConstraint.containerIdentity ||
-					canonicalBytes(relation.references) !==
-						canonicalBytes(constraint.fields)
-				)
-					continue;
-
-				const sourceIdentity = String(sourceCollection.identity);
-				const targetSourceIdentity = mapIdentityForward(
-					sourceIdentity,
-					renames,
-				);
-				const targetSource = targetCollections.get(targetSourceIdentity);
-				if (!targetSource) {
-					const dropCollection = stepByKindAndIdentity.get(
-						`dropCollection\0${sourceIdentity}`,
-					);
-					if (dropCollection) depend(dropCollection, droppedConstraint);
-					continue;
-				}
-
-				const baseRelationIdentity = String(relation.identity);
-				const targetRelationIdentity = mapIdentityForward(
-					baseRelationIdentity,
-					renames,
-				);
-				const dropRelation = ensureStep(
-					step({
-						kind: "dropRelation",
-						targetIdentity: baseRelationIdentity,
-						containerIdentity: sourceIdentity,
-						lock: "accessExclusive",
-						scansData: true,
-						rewritesTable: false,
-						reversibleWithoutData: false,
-						classification: "destructive",
-					}),
-				);
-				depend(dropRelation, droppedConstraint);
-
-				const targetRelation = childRecords(targetSource, "relations").find(
-					(candidate) => candidate.identity === targetRelationIdentity,
-				);
-				if (!targetRelation) continue;
-				const addRelation = ensureStep(
-					step({
-						kind: "addRelation",
-						targetIdentity: targetRelationIdentity,
-						containerIdentity: targetSourceIdentity,
-						lock: "accessExclusive",
-						scansData: true,
-						rewritesTable: false,
-						reversibleWithoutData: false,
-						classification: "destructive",
-					}),
-				);
-				depend(droppedConstraint, addRelation);
-				if (replacementConstraint) depend(replacementConstraint, addRelation);
-				const referencedTarget = targetCollections.get(
-					String(targetRelation.target),
-				);
-				for (const targetConstraint of referencedTarget
-					? childRecords(referencedTarget, "constraints")
-					: []) {
-					if (
-						(targetConstraint.kind === "primaryKey" ||
-							targetConstraint.kind === "unique") &&
-						canonicalBytes(targetConstraint.fields) ===
-							canonicalBytes(targetRelation.references)
-					) {
-						const addedKey = stepByKindAndIdentity.get(
-							`addConstraint\0${String(targetConstraint.identity)}`,
-						);
-						if (addedKey) depend(addedKey, addRelation);
-					}
-				}
-			}
-		}
-	}
-	for (const addedRelation of steps.filter(
-		(candidate) => candidate.kind === "addRelation",
-	)) {
-		const sourceCollection = targetCollections.get(
-			addedRelation.containerIdentity,
-		);
-		const relation = sourceCollection
-			? childRecords(sourceCollection, "relations").find(
-					(candidate) => candidate.identity === addedRelation.targetIdentity,
-				)
-			: undefined;
-		if (!relation) continue;
-		const referencedTarget = targetCollections.get(String(relation.target));
-		for (const targetConstraint of referencedTarget
-			? childRecords(referencedTarget, "constraints")
-			: []) {
-			if (
-				(targetConstraint.kind === "primaryKey" ||
-					targetConstraint.kind === "unique") &&
-				canonicalBytes(targetConstraint.fields) ===
-					canonicalBytes(relation.references)
-			) {
-				const addedKey = stepByKindAndIdentity.get(
-					`addConstraint\0${String(targetConstraint.identity)}`,
-				);
-				if (addedKey) depend(addedKey, addedRelation);
-			}
-		}
-		for (const droppedKey of steps.filter(
-			(candidate) => candidate.kind === "dropConstraint",
-		)) {
-			if (
-				mapIdentityForward(droppedKey.containerIdentity, renames) !==
-				relation.target
-			)
-				continue;
-			const baseTarget = baseCollections.get(droppedKey.containerIdentity);
-			const baseConstraint = baseTarget
-				? childRecords(baseTarget, "constraints").find(
-						(candidate) => candidate.identity === droppedKey.targetIdentity,
-					)
-				: undefined;
-			if (
-				baseConstraint &&
-				(baseConstraint.kind === "primaryKey" ||
-					baseConstraint.kind === "unique") &&
-				canonicalBytes(semanticComparable(baseConstraint.fields, renames)) ===
-					canonicalBytes(relation.references)
-			)
-				depend(droppedKey, addedRelation);
-		}
-	}
-	return sortSteps(steps, renames, dependencies);
+	const expanded = expandReferencedKeyDependencies({
+		base,
+		target,
+		renames,
+		steps,
+	});
+	return sortSteps(expanded.steps, renames, expanded.dependencies);
 }
